+++ cpp/l2-proxy/l2-proxy-version.h
@@ -1,4 +1,4 @@
 #ifndef L2_PROXY_VERSION_H
 #define L2_PROXY_VERSION_H
-#define VERSION "1.0.4-3259eeb"
+#define VERSION "1.0.4-b7e8780"
 #endif